pyproject.toml
[tool.poetry]
name = "twitter-trending-archive-api"
version = "0.1.0"
description = "Twitter Trending Archive API"
authors = ["Logan Cox <mail@logan-cox.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "twitter_trending_archive_api"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
apache-airflow = "^2.5.1"
tweepy = "^4.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
